+++ src/main/java/io/contek/invoker/mexc/spot/api/common/_OrderBookLevel.java
@@ -1,9 +1,9 @@
 package io.contek.invoker.mexc.spot.api.common;
 
 import javax.annotation.concurrent.NotThreadSafe;
-import java.math.BigDecimal;
+import java.math.Double;
 import java.util.ArrayList;
 
 @NotThreadSafe
-public class _OrderBookLevel extends ArrayList<BigDecimal> {
+public class _OrderBookLevel extends ArrayList<Double> {
 }
